hack: introduce temporary fix to make simulations run
The variable name `method` is in conflict. It's being used for things
like "PELT" and "loess" within the `demean` and `desmooth` functions in
addition to correlation applications like "ECE" "desmooth" and "demean".

The problem is especially pronounced because we add the ellipses for
additional arguments, `method` gets automatically interpretd as an
additional argument rather than a core argument.